refactor: move device flow module to github/
